Moved around tests so they would be in the correct folders
